api/requirements.txt
beautifulsoup4==4.9.3
django-environ==0.4.5
Flask==2.0.0
lxml==4.6.3
requests==2.25.1
virtualenv==20.4.5
